Ej herencia restaurante

+++ src/UT4/herencia/ejHerencias3/Pescados.java
@@ -2,7 +2,7 @@ package UT4.herencia.ejHerencias3;
 
 public class Pescados extends Platos{
 
-    public Pescados(String nombre, int precio) {
+    public Pescados(String nombre, double precio) {
         super(nombre, precio);
     }
